Extract reduce_scatter_common shared utilities for reduce scatter
The next commit will implement strided reduce scatter whose
implementation is largely based on minimal
reduce scatter. Extract out the common part to avoid code
repetition.
